HSEARCH-2878 Move analyzer-related tests to the -engine module
Those tests have nothing to do with the ORM integration and they caused
encoding-related problems in the MySQL ITs.
