+++ engine/src/test/java/org/hibernate/search/test/analyzer/common/AnalyzerForTests1.java
@@ -4,21 +4,23 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.test.analyzer;
+package org.hibernate.search.test.analyzer.common;
 
 import org.apache.lucene.util.Version;
 
 /**
  * @author Emmanuel Bernard
  */
-public final class AnalyzerForTests2 extends AbstractTestAnalyzer {
-	private final String[] tokens = { "sound", "cat", "speed" };
+public final class AnalyzerForTests1 extends AbstractTestAnalyzer {
 
-	public AnalyzerForTests2(Version version) {
+	private final String[] tokens = { "alarm", "dog", "performance" };
+
+	public AnalyzerForTests1(Version version) {
 	}
 
 	@Override
 	protected String[] getTokens() {
 		return tokens;
 	}
+
 }
