+++ engine/src/test/java/org/hibernate/search/test/analyzer/common/BlogEntry.java
@@ -4,17 +4,17 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.test.analyzer;
+package org.hibernate.search.test.analyzer.common;
 
 import java.util.Set;
 
 import org.hibernate.search.annotations.AnalyzerDef;
+import org.hibernate.search.annotations.AnalyzerDefs;
 import org.hibernate.search.annotations.AnalyzerDiscriminator;
 import org.hibernate.search.annotations.DocumentId;
 import org.hibernate.search.annotations.Field;
 import org.hibernate.search.annotations.Indexed;
 import org.hibernate.search.annotations.IndexedEmbedded;
-import org.hibernate.search.annotations.Parameter;
 import org.hibernate.search.annotations.Store;
 import org.hibernate.search.annotations.TokenFilterDef;
 import org.hibernate.search.annotations.TokenizerDef;
@@ -28,26 +28,28 @@ import org.apache.lucene.analysis.standard.StandardTokenizerFactory;
  * @author Hardy Ferentschik
  */
 @Indexed
-@AnalyzerDef(name = "en",
-		tokenizer = @TokenizerDef(factory = StandardTokenizerFactory.class),
-		filters = {
-				@TokenFilterDef(factory = LowerCaseFilterFactory.class),
-				@TokenFilterDef(factory = SnowballPorterFilterFactory.class, params = {
-					@Parameter(name = "language", value = "English")
+@AnalyzerDefs({
+		@AnalyzerDef(name = "en",
+				tokenizer = @TokenizerDef(factory = StandardTokenizerFactory.class),
+				filters = {
+						@TokenFilterDef(factory = LowerCaseFilterFactory.class),
+						@TokenFilterDef(factory = SnowballPorterFilterFactory.class
+						)
+				}),
+		@AnalyzerDef(name = "de",
+				tokenizer = @TokenizerDef(factory = StandardTokenizerFactory.class),
+				filters = {
+						@TokenFilterDef(factory = LowerCaseFilterFactory.class),
+						@TokenFilterDef(factory = GermanStemFilterFactory.class)
 				})
-		})
-@AnalyzerDef(name = "de",
-		tokenizer = @TokenizerDef(factory = StandardTokenizerFactory.class),
-		filters = {
-				@TokenFilterDef(factory = LowerCaseFilterFactory.class),
-				@TokenFilterDef(factory = GermanStemFilterFactory.class)
-		})
-public class Article {
+})
+@AnalyzerDiscriminator(impl = LanguageDiscriminator.class)
+public class BlogEntry {
 
 	private Integer id;
 	private String language;
 	private String text;
-	private Set<Article> references;
+	private Set<BlogEntry> references;
 
 	@DocumentId
 	public Integer getId() {
@@ -78,11 +80,11 @@ public class Article {
 	}
 
 	@IndexedEmbedded(depth = 1)
-	public Set<Article> getReferences() {
+	public Set<BlogEntry> getReferences() {
 		return references;
 	}
 
-	public void setReferences(Set<Article> references) {
+	public void setReferences(Set<BlogEntry> references) {
 		this.references = references;
 	}
 }
