+++ engine/src/test/java/org/hibernate/search/test/analyzer/common/DuplicatedAnalyzerDefinitionTest.java
@@ -5,7 +5,7 @@
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
 
-package org.hibernate.search.test.analyzer;
+package org.hibernate.search.test.analyzer.common;
 
 import java.lang.annotation.ElementType;
