+++ engine/src/test/java/org/hibernate/search/test/analyzer/definition/AnalyzerBuilderTest.java
