+++ engine/src/test/java/org/hibernate/search/test/analyzer/definition/InsertWhitespaceFilterFactory.java
