+++ engine/src/test/java/org/hibernate/search/test/analyzer/definition/Team.java
