+++ engine/src/test/java/org/hibernate/search/test/analyzer/inheritance/AnalyzerInheritanceTest.java
