+++ engine/src/test/java/org/hibernate/search/test/analyzer/inheritance/BaseClass.java
