+++ engine/src/test/resources/org/hibernate/search/test/analyzer/mapping-chars.properties
