+++ engine/src/test/resources/org/hibernate/search/test/analyzer/stemmer-override.properties
