+++ engine/src/test/resources/org/hibernate/search/test/analyzer/stoplist.properties
