+++ engine/src/test/resources/org/hibernate/search/test/analyzer/synonyms.properties
